move auth.ts to lib
