+++ packages/server/lib/auth.ts
